+++ application/controllers/Payments.php
@@ -12,7 +12,7 @@ class Payments extends MY_Controller
         $this->load->view('pages/payments/list');
     }
 
-     /**
+    /**
      * Show a resource
      * html view
      */
@@ -24,7 +24,7 @@ class Payments extends MY_Controller
         $this->load->view('pages/payments/detail', $data);
     }
 
-     /**
+    /**
      * Show a form page for creating resource
      * html view
      */
@@ -33,7 +33,7 @@ class Payments extends MY_Controller
         $this->load->view('pages/payments/edit');
     }
 
-     /**
+    /**
      * Show a form page for updating resource
      * html view
      */
@@ -49,23 +49,31 @@ class Payments extends MY_Controller
      * Store a resource
      * print json Response
      */
-    public function store ()
+    public function store()
     {
-        $record = $this->input->post();
-        $repayment  = $this->payment->create($record);
-        $error = $this->session->flashdata('error_message') . $this->session->flashdata('warning_message');
+        $gate = auth()->can('create', 'payment');
+        if ($gate->allowed()) {
+            $record = $this->input->post();
+            $repayment  = $this->payment->create($record);
+            $error = $this->session->flashdata('error_message') . $this->session->flashdata('warning_message');
 
-        if ($repayment) {
-            $out = [
-                'data' => $repayment,
-                'input' => $record,
-                'status' => true,
-                'message' => 'Repayment made successfully!'
-            ];
+            if ($repayment) {
+                $out = [
+                    'data' => $repayment,
+                    'input' => $record,
+                    'status' => true,
+                    'message' => 'Repayment made successfully!'
+                ];
+            } else {
+                $out = [
+                    'status' => false,
+                    'message' => $error ? $error : "Data couldn't be processed!"
+                ];
+            }
         } else {
             $out = [
                 'status' => false,
-                'message' => $error?$error:"Data couldn't be processed!"
+                'message' => $gate->message
             ];
         }
         httpResponseJson($out);
@@ -75,24 +83,31 @@ class Payments extends MY_Controller
      * Update a resource
      * print json Response
      */
-    public function update (int $id = null)
+    public function update(int $id = null)
     {
-        $record = $this->input->post();
-        $loan_payment  = $this->payment->update($id, $record);
-        $error = $this->session->flashdata('error_message') . $this->session->flashdata('warning_message');
+        $gate = auth()->can('update', 'payment', $this->payment->find($id));
+        if ($gate->allowed()) {
+            $record = $this->input->post();
+            $loan_payment  = $this->payment->update($id, $record);
+            $error = $this->session->flashdata('error_message') . $this->session->flashdata('warning_message');
 
-        if($loan_payment){
-            $out = [
-                'data' => $loan_payment,
-                'input' => $record,
-                'status' => true,
-                'message' => 'Loan payment data updated successfully!'
-            ];
-        }
-        else {
+            if ($loan_payment) {
+                $out = [
+                    'data' => $loan_payment,
+                    'input' => $record,
+                    'status' => true,
+                    'message' => 'Loan payment data updated successfully!'
+                ];
+            } else {
+                $out = [
+                    'status' => false,
+                    'message' => $error ? $error : "Loan payment data couldn't be update!"
+                ];
+            }
+        } else {
             $out = [
                 'status' => false,
-                'message' => $error?$error:"Loan payment data couldn't be update!"
+                'message' => $gate->message
             ];
         }
         httpResponseJson($out);
@@ -102,24 +117,30 @@ class Payments extends MY_Controller
      * Delete a resource
      * print json Response
      */
-    public function delete (int $id = null)
+    public function delete(int $id = null)
     {
-
-        if($this->payment->delete($id)){
-            $out = [
-                'status' => true,
-                'message' => 'Loan repayment deleted successfully!'
-            ];
-        }
-        else {
+        $gate = auth()->can('delete', 'payment', $this->payment->find($id));
+        if ($gate->allowed()) {
+            if ($this->payment->delete($id)) {
+                $out = [
+                    'status' => true,
+                    'message' => 'Loan repayment deleted successfully!'
+                ];
+            } else {
+                $out = [
+                    'status' => false,
+                    'message' => "Loan repayment couldn't be deleted!"
+                ];
+            }
+        } else {
             $out = [
                 'status' => false,
-                'message' => "Loan repayment couldn't be deleted!"
+                'message' => $gate->message
             ];
         }
         httpResponseJson($out);
     }
-   
+
     public function datatables()
     {
         $start = $this->input->get('start', true);
@@ -127,10 +148,10 @@ class Payments extends MY_Controller
         $draw = $this->input->get('draw', true);
         $inputs = $this->input->get();
         $query = $this->payment->all();
-        if($this->input->get('loan_id')){
+        if ($this->input->get('loan_id')) {
             $query->where(['loan_id' => $inputs['loan_id']]);
-       }
-        
+        }
+
         $out = datatable($query, $start, $length, $draw, $inputs);
         $out = array_merge($out, [
             'input' => $this->input->get(),
